flog: transfer files from separate subtree to criu/criu
Delete the criu/flog and move its files into criu/criu.
Initially, flog had been added as a separate subtree in
criu. This causes a hindrance while using functions defined
in criu into the files of flog.
