+++ criu/include/flog_util.h
