--- flog/built-in.S
@@ -1,4 +1,0 @@
-SECTIONS
-{
-	.rodata : { _rodata_start = . ; *(.rodata*) ; _rodata_end = . ;}
-}
