--- flog/include/types.h
@@ -1,16 +1,0 @@
-#ifndef __FLOG_TYPES_H__
-#define __FLOG_TYPES_H__
-
-#include <stdint.h>
-#include <stdbool.h>
-
-typedef uint64_t	u64;
-typedef int64_t		s64;
-typedef uint32_t	u32;
-typedef int32_t		s32;
-typedef uint16_t	u16;
-typedef int16_t		s16;
-typedef uint8_t		u8;
-typedef int8_t		s8;
-
-#endif /* __FLOG_TYPES_H__ */
