--- flog/src/Makefile
@@ -1,5 +1,0 @@
-ccflags-y += -DCONFIG_X86_64 -iquote ./include $(OPTS)
-ldflags-y		+= -r
-
-#obj-y += main.o
-obj-y += flog.o
